go/ssa: build methods of RuntimeTypes lazily
This change causes the builder to no longer eagerly create the
complete method set for each type in Program.RuntimeTypes.
Instead, methods are created on demand, such as during the
traversal performed by ssautil.AllFunctions.
We expect this to reduce the number of wrappers created in
single-package scenarios such as go/analysis/passes/buildssa.

Each time the builder encounters a MakeInterface whose operand is a
non-interface ground type, it adds the type, along with all types
derivable from it via reflection, to the set Program.runtimeTypes.
This set is guarded by a dedicated mutex.
(The derivation algorithm, forEachReachable, is abstracted from the
builder and some mistakes re: tuples and signatures are corrected. We
may want to share it with go/callgraph/rta.)

Since an encounter of a new runtime type no longer causes creation of
methods, we can simplify the builder iteration substantially.

For now, we preserve the existing behavior of RuntimeTypes,
which requires two kludges:
1. the existing kludge in Package.Build to add types of exported
   package members;
2. a kludge added to RuntimeTypes to subtract interface types
   and types without methods.
We will remove both kludges in the following change,
which will clarify and simplify the definition of RuntimeTypes.

Also, remove two dead functions.

Change-Id: Ied157d08096e9f8c92671c266738e0ba26ba1acb
Reviewed-on: https://go-review.googlesource.com/c/tools/+/538358
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Tim King <taking@google.com>