go/ssa: create methods on demand
This change enables the SSA builder to create all
methods on demand as the need (e.g. a call) is encountered.

This allows package-at-a-time analyses to call CreatePackage
only for the direct imports of a syntax package, but
not for all the indirect dependencies as well, which
improves the asmyptote for separate analysis (e.g. in
unitchecker or gopls).

Details:
- Prog.objectMethod creates a method from a types.Func,
  even if the package was not created.
  It combines the work of
  originFunc (RIP) + receiverTypeArgs + needsInstance;
  this triad of calls used to appear three times.
- Delete originFunc declaredFunc.
- Doc tweaks.
- Unit test

Also:
- go/analysis/passes/buildssa: don't call CreatePackage
  for indirect dependencies.

Change-Id: I905ea6e52473a4fe6a73ed574ea0f0ce0e8fe7ed
Reviewed-on: https://go-review.googlesource.com/c/tools/+/538763
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Tim King <taking@google.com>
Auto-Submit: Alan Donovan <adonovan@google.com>