all: replace all usages of golang.org/x/sys/execabs with os/exec
The "golang.org/x/sys/execabs" package was introduced to address a
security issue on Windows, and changing the default behavior of os/exec
was considered a breaking change. go1.19 applied the behavior that was
previously implemented in the execabs package through CL 403274.

This reverts fe37c9e135b934191089b245ac29325091462508 (CL 284773).
Updates #43724.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>