go/ssa: remove Function.Syntax optimization
As an optimization, Function.Syntax didn't guarantee
to return the actual syntax for a function in all cases,
and would instead return a stub that provides Pos/End.
The aim was to prevent SSA from keeping ASTs live,
but in practice that always happens anyway; and it
had the perverse effect of forcing users to set
GlobalDebug to get the syntax, which is even more
expensive in space.

This change retains the (syntax, info, goversion) triad
unconditionally.

This enables a simplification of the Function.generic,
which no longer needs retain them.

Change-Id: I41dd22a995d1093d2ecfd9cca743b7606bd5cc8b
Reviewed-on: https://go-review.googlesource.com/c/tools/+/539257
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Tim King <taking@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>