gopls/internal/lsp/cache: allow duplicate analyzer names
The Actions map is now keyed by an analyzer's stable name,
which has a suffix consisting of the code address of its Run
function. This allows us to use two analyzers called
nilness in the same action graph, so long as their implementations
are different.

(Two analyzers with the same Name and Run code--e.g. closures
with different values--is a theoretical but remote possibility.)

Change-Id: Iecc21a61f666a1f6d1a71c8691c31ec3fd039875
Reviewed-on: https://go-review.googlesource.com/c/tools/+/539495
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>