gopls/internal/regtest/marker: porting extract tests
Port the extraction tests, which involved a mixture of suggestedfix,
extractmethod, and extractfunc markers (notably, nothing was running the
extractfunc markers!). To do this, extend the new codeaction markers to
accept an option slice of titles to use for filtering.

Modify the method/func extraction tests to reduce verbosity and long
lines.

For golang/go#54845

Change-Id: I319d1c731a4cdb4ad952944667bc57d781ad176d
Reviewed-on: https://go-review.googlesource.com/c/tools/+/539481
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>