gopls/internal/lsp/cache: remove forceReloadMetadata from clone
We need to simplify Clone for zero config gopls. One source of
complexity is the 'forceReloadMetadata' parameter, which reaches inside
the clone internals to force invalidation of metadata, and must be
threaded through the call stack, originating from the
'InvalidateMetadata' action on a file.

Instead, just recreate the View, which will have the same side effect of
invoking `go list`.

As a nice side-effect, we can actually use the FromRegenerateCgo
diagnostic source, which allows us to avoid the unbounded Await in
TestRegenerateCgo.

For golang/go#57979

Change-Id: I7fa01ed17a9407dfb98844eeec69b1fdbbd27aa7
Reviewed-on: https://go-review.googlesource.com/c/tools/+/539658
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>