gopls/internal/golang: "Show free symbols" code action
This change adds a new "Show free symbols" (source.freesymbols)
code action that reports the set of free symbols referenced by
the selected region of Go source. The HTML report, produced by
the /freerefs endpoint of gopls' web server, includes an
itemized list of symbols, and a color-annotated source listing.

Symbols are presented in three groups: imported symbols
(grouped by package); package-level symbols and local symbols.
Each symbol is a link to either the integrated doc viewer
(for imported symbols) or the declaration, for others.

The feature is visible in editors as:
- VS Code: Source actions... > Show free references
- Emacs+eglot: M-x go-freesymbols
  (Requires (eglot--code-action go-freerefs "source.freesymbols")
   until dominikh/go-mode.el#436 is resolved.)

There are a number of opportunities for factoring in common
with RenderPackageDoc; they will be dealt with in a follow-up.

Also:
- a unit test of the freeRefs algorithm;
- an integration test of the web interaction.
- release notes.

Change-Id: I97de76686fcc28e445a72e7c611673c47e467dfd
Reviewed-on: https://go-review.googlesource.com/c/tools/+/539663
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>