gopls/internal/lsp: factor showMessage calls
Also, decouple a number of places where the success or
failure to show a message was mistakenly being reported
as the success or failure of the enclosing operation.
Because the new helper function has no result, this
temptation is gone.

Also, made addFolders no longer return an error.
Previously, the failure of showMessage would cause
addFolders (and Did{Open,ChangeWatchedFolders}) to
fail for a rather trivial reason.

Change-Id: Ia08815f2a4a28b98832b422213e2eb3001772704
Reviewed-on: https://go-review.googlesource.com/c/tools/+/541124
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>