gopls/internal/lsp/tests: remove AddImport support
I went to translate the one 'addimport' marker test to the new
framework, only to decide that it would be better as a one-off
regression test in ./internal/regtest/misc, only to discover that there
was already a regression test misc.TestAddImport doing exactly what we
want. So this coverage was completely redundant -- just remove it.

For golang/go#54845

Change-Id: I1412990a9027b97b2b76c56462fe17c43c5a0c6e
Reviewed-on: https://go-review.googlesource.com/c/tools/+/541126
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>