gopls/internal/lsp/cache: move go.work parsing into the view
For golang/go#57979, the set of modules used by a go.work file will be
an input into the View calculation. Therefore, it can't be tracked
inside the snapshot, and must be computed outside of the view.

Implement this, and use go.work parsing to detect view changes.

For golang/go#57979

Change-Id: I40964a6b4f7194740f455cf2986d1d37ae3b522f
Reviewed-on: https://go-review.googlesource.com/c/tools/+/542619
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>