gopls/internal/lsp: eliminate server code generation
The "helper" program generated "not implemented" stubs
for each server method in the protocol, presumably so
that upgrading the LSP would not break the build when
new methods are added. The value seems rather marginal:
knowing that the interface has expanded is useful, and
it's easy to address manually.

Change-Id: I4fd5d9baa9056c5f62fd6ff8dfd042053f0b1c41
Reviewed-on: https://go-review.googlesource.com/c/tools/+/543556
Auto-Submit: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>