gopls/internal/lsp/source: fix hover for objects in unsafe
Refactor to fix a panic during hover on objects in unsafe, using the
recently refined decl logic for definition.

For the new test, matching the unsafe package documentation exactly is
liable to be a change detector with the standard library, so refactor
the marker test to support more flexible string matching for hover,
reusing logic from the wantError type.

While simplifying, eliminate markdown normalization in tests, as it is
only needed for testing Go 1.18, which is soon to be deprecated anyway.

Fixes golang/go#64239

Change-Id: I5fd986b6db1dd77502f78fd88ad4f5c7a44b48bb
Reviewed-on: https://go-review.googlesource.com/c/tools/+/544616
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>