gopls/internal/lsp: invert the import between cache and source
After much pulling at this thread, finally invert the relationship
between the cache package and source package.

In order to do this, the implementation of the mod vuln promise had to
be moved from the internal/vulncheck/scan package into the cache package
(to avoid cycles). This is consistent with the implementation of other
promises in the cache package.

Change-Id: I34c9069231dd27a4fc0d88ae9d688d2a38e81d27
Reviewed-on: https://go-review.googlesource.com/c/tools/+/545315
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>