gopls/internal/regtest: avoid file segments ending in a period
The telemetry test used the response string "Yes, ..." as
a subtest name, causing the testing package to create a
temp dir with this string as a path segment. Apparently,
Windows does not allow a path segment to end in a period,
and gopls checks for this in checkPathCase (now renamed
to checkPathValid to reflect its broader nature).
This caused addFolders to fail, leading to extremely
confusing failures in recent CLs.

Also, print a warning in regtest if a test tempdir
would be an invalid workspace folder dir, to spoil
similar mysteries in future.

Change-Id: Id6368cb9718bf997215bcb3b8bfaeeb6a7505c9d
Reviewed-on: https://go-review.googlesource.com/c/tools/+/545595
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>