gopls/internal/server: rewrite the server diagnostic tracking
For zero-config gopls, we need diagnostics to work seamlessly across
multiple views. The existing diagnostic logic is already complicated, so
rewrite it to significantly simplify and allow for merging diagnostics
from multiple views. Rather than using a global snapshot ID as a counter
that enforces freshness, use the set of views and their snapshot
sequence IDs.

Replace the store->publish model with a simple update model, so that we
store and publish in one atomic operation. I believe this avoids most of
the races that the older more complicated schema was designed around.

For simplicity, delete the formatting of diagnostics in debug templates.
I do not think they provide any value, and so there is no point in
maintaining them.

For golang/go#57979

Change-Id: I4c595c1d7c8b4f08fce490e2a49b8a6aece41e92
Reviewed-on: https://go-review.googlesource.com/c/tools/+/546415
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>