gopls/internal/lsp/protocol/generate: restore indentation
CL 545015 tidied the source-level indentation of the
DocumentURI declaration in the generator, but this
added unwanted indentation in the output (which, for
reasons I still haven't investigated, was not fixed
by the final "go fmt" operation done by the generator).

This change restores the old indentation so that the
output is properly formatted.

Change-Id: Iaa3c0cc3930218f6deed3f6de21dde5a0019aaa8
Reviewed-on: https://go-review.googlesource.com/c/tools/+/546657
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Peter Weinberger <pjw@google.com>