gopls/internal/analysis: improve analyzer docs
This change moves MustExtractDoc to a higher (but still internal)
directory so that it can be used from gopls, and converts all
of gopls' analyzers into the form it expects.

It also adds the corresponding pkg.go.dev page to
each Analyzer.URL, and shows this link in the generated markdown.

Even if ExtractDoc is never published (see golang/go#61315) it
is still convenient to use in x/tools.

Fixes golang/go#63820

Change-Id: Ib3e047bef591574154d5656db69e1609aaf30a4a
Reviewed-on: https://go-review.googlesource.com/c/tools/+/547877
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>