gopls/internal/server: rewrite the server diagnostic tracking
For zero-config gopls, we need diagnostics to work seamlessly across
multiple views. The existing diagnostic logic is already complicated, so
rewrite it to significantly simplify and allow for merging diagnostics
from multiple views. Rather than using a global snapshot ID as a counter
that enforces freshness, use the set of views and their snapshot
sequence IDs.

Replace the store->publish model with a simple update model, so that we
store and publish in one atomic operation. I believe this avoids most of
the races that the older more complicated schema was designed around.

For simplicity, delete the formatting of diagnostics in debug templates.
I do not think they provide any value, and so there is no point in
maintaining them.

Note that this reverts CL 546855, rolling forward CL 546415, As the
deadlock was fixed in CL 549415.

Change-Id: Ia5b63c4ee87ec29fab334badd81989d0f0f260b5
Reviewed-on: https://go-review.googlesource.com/c/tools/+/549895
Reviewed-by: Bryan Mills <bcmills@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>