gopls/internal/lsp/cache: add the zero-config algorithm, and a unit test
Add an implementation of the new zero-config workspace algorithm, by way
of new logic for bestViewForURI, a forURI argument for getViewDefinition
(renamed to defineView), and a new selectViews algorithm.

For now, this logic is only exercised in a new unit test. Subsequent CLs
will refine this logic, and wire it in to gopls. Once this is wired in a
lot of tests will need to change, so it is probably simpler to first
review the logic in isolation, as is done here.

For golang/go#57979

Change-Id: I150ba87ac9e51119f35e095b4c41984b5c4f9b47
Reviewed-on: https://go-review.googlesource.com/c/tools/+/550815
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>