gopls/internal/lsp/cache: finish integrating the zero-config algorithm
This CL finishes wiring in the zero-config gopls algorithm of
golang/go#57979, by using the new logic for updating the set of views.

Most of the heavy lifting was done in previous CLs, so the change here
is relatively trivial. Tests had to be adjusted to account for the lack
of diagnostics now that the zero-config logic allows module loading to
succeed in more cases.

For golang/go#57979

Change-Id: I98aeea8777a354f22c952875b0e655c535d4b995
Reviewed-on: https://go-review.googlesource.com/c/tools/+/551896
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>