gopls/internal/lsp/source: simplify SuggestedFix plumbing
This change simplifies the interface to "suggested fix"
functions so that the general case and single-file case
are more alike.

The argument types have changed thus:
(file.Handle, protocol.Range) ->
 (pkg *cache.Package, pgf *parsego.File, start, end token.Pos)
and the results thus:
 []protocol.TextDocumentEdit -> (*token.FileSet, *analysis.SuggestedFix)

The extra FileSet result is needed because Pos values in
the SuggestedFix are not necessarily mapped by pkg.FileSet().

The logic of the singleFile wrapper moved into ApplyFix,
so it is now trivial, but unfortunately we can't get rid
of it yet because it breaks an import cycle from the
analyzer packages to cache. We first need to stop using
analyzers for quick fixes.

Also:
- rename suggestedFixFunc -> Fixer
- add various missing comments.
- merge the two parts of stubMethodsFixer.

(This was preparatory work while trying to understand
the associated issue.)

Updates golang/vscode-go#3101

Change-Id: I4602f7e22730a9037707e13f08b52d5cb7c714d6
Reviewed-on: https://go-review.googlesource.com/c/tools/+/552735
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Alan Donovan <adonovan@google.com>