go/packages: wait for sizes goroutine on all control paths
This change ensures that the sizes goroutine created by
goListDriver always terminates before the latter function
returns. We suspect that the failure to wait in some of
the early-return error cases causes the child process
to retain a references to the file system that (on Windows)
prevented garbage collection of the test's temporary
directory.

Fixes golang/go#65180

Change-Id: I755ed8956df338e0a53bf08a7451e4e62d08bafc
Reviewed-on: https://go-review.googlesource.com/c/tools/+/560478
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>
Reviewed-by: Michael Matloob <matloob@golang.org>